Updated model.

+++ src/polly/database/existence.go
@@ -7,7 +7,9 @@ import (
 )
 
 // TODO check when this error occcurs, maybe we could just return a bool
-func (db *Database) ExistsParticipant(userID int64, pollID int) (bool, error) {
+func (db *Database) ExistsParticipant(userID int64, pollID int64) (bool,
+	error) {
+
 	count, err := db.mapping.SelectInt(fmt.Sprintf(
 		"select count(1) from %s where %s=$1 and %s=$2;",
 		cParticipantTableName, cUserID, cPollID), userID, pollID)
